Fixes #3190: Fix custom field rendering for Jinja2 export templates
